Split out the worker_context code, also drop the list of listeners from the worker context
